Integrate ESBuild, refactored to only use esbuild for all building purposes instead of webpack
