+++ lib/v1/css/texscript-core.css
@@ -5,6 +5,7 @@
   box-sizing: border-box;
   box-shadow: 4px 8px 16px rgba(0, 0, 0, 0.5);
   page-break-after: always;
+  overflow: clip;
 }
 
 .texscript-Line {
