--- lib/v2/css/styles.css
@@ -1,5 +1,0 @@
-@import "./base.css";
-@import "./print.css";
-@import "./texscript-core.css";
-@import "./texscript-List.css";
-@import "./texscript-Page.css";
