+++ lib/v2/css/texscript.css
@@ -1,0 +1,5 @@
+@import "./native/base.css";
+@import "./media/print.css";
+@import "./custom/texscript-core.css";
+@import "./custom/texscript-List.css";
+@import "./custom/texscript-Page.css";
