+++ lib/v2/texscript.lib.ts
@@ -1,0 +1,4 @@
+export async function load() {
+  const splash = await import("./lib/splash.js");
+  await splash.loadSplash();
+}
