+++ lib/v2/types/global.d.ts
@@ -6,6 +6,7 @@ declare global {
       version: string;
       repoURL: string;
       lastCompilation: Map<string, unknown>;
+      toggleSplash: Function;
     };
   }
 }
