+++ package.json
@@ -6,40 +6,31 @@
     "build": "npm run v1-build && npm run v2-build",
     "serve": "node servers/app.js",
     "start": "npm run serve",
-    "v1-build": "npx rimraf build/texscript/v1 && npx webpack --config scripts/v1/webpack.config.js --env outputDir=build mode=production",
-    "v1-build-dev": "npx rimraf build/texscript/v1 && npx webpack --config scripts/v1/webpack.config.js",
-    "v1-serve": "node servers/v1/lib-watcher-builder.js",
+    "v1-build": "npm run v1-build:bundle",
+    "v1-build:bundle": "node scripts/v1/build-bundle.js",
+    "v1-dev:all": "npm run v1-dev:bundle",
+    "v1-dev:bundle": "node scripts/v1/dev-bundle.js",
     "v1-preview": "npm run v1-build && npx http-server build --port=80",
-    "v2-build": "npx rimraf build/texscript/v2 && npx webpack --config scripts/v2/webpack.config.js --env outputDir=build mode=production",
-    "v2-build-dev": "npx rimraf build/texscript/v2 && npx webpack --config scripts/v2/webpack.config.js",
-    "v2-serve": "node servers/v2/lib-watcher-builder.js",
-    "v2-preview": "npm run v2-build && npx http-server build --port=80"
+    "v1-serve": "node servers/v1/lib-watcher-builder.js",
+    "v2-build": "npm run v2-build:bundle",
+    "v2-build:bundle": "node scripts/v2/build-bundle.js",
+    "v2-dev:all": "npm run v2-dev:bundle",
+    "v2-dev:bundle": "node scripts/v2/dev-bundle.js",
+    "v2-preview": "npm run v2-build && npx http-server build --port=80",
+    "v2-serve": "node servers/v2/lib-watcher-builder.js"
   },
   "author": "",
   "license": "ISC",
   "description": "",
   "type": "module",
   "devDependencies": {
-    "@babel/core": "^7.28.4",
-    "@babel/preset-env": "^7.28.3",
-    "@babel/preset-typescript": "^7.27.1",
-    "babel-loader": "^10.0.0",
     "chokidar": "^4.0.3",
-    "copy-webpack-plugin": "^12.0.2",
-    "css-loader": "^7.1.2",
-    "css-minimizer-webpack-plugin": "^7.0.2",
     "depcheck": "^1.4.7",
+    "esbuild": "^0.25.10",
     "express": "^5.1.0",
     "http-proxy-middleware": "^3.0.5",
     "http-server": "^14.1.1",
-    "ignore-emit-webpack-plugin": "^2.0.6",
-    "mini-css-extract-plugin": "^2.9.4",
     "serve-index": "^1.9.1",
-    "terser-webpack-plugin": "^5.3.14",
-    "ts-loader": "^9.5.4",
-    "typescript": "^5.9.3",
-    "webpack": "^5.102.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-merge": "^6.0.1"
+    "typescript": "^5.9.3"
   }
 }
